+++ rust/template/ddlog_derive/tests/ui/pass/mutator.rs
@@ -1,5 +1,9 @@
 use ddlog_derive::{FromRecord, IntoRecord, Mutator};
 use serde::Deserialize;
+use std::{
+    collections::BTreeMap,
+    vec::Vec,
+};
 
 fn main() {}
 
@@ -20,7 +24,7 @@ struct UnitStruct;
 #[derive(FromRecord, IntoRecord, Mutator, Debug, Clone, Default, PartialEq, Deserialize)]
 struct TupleStruct(u32, u64, Vec<u8>, String);
 
-#[derive(FromRecord, IntoRecord, Mutator, Debug, Clone, PartialEq, Deserialize)]
+#[derive(FromRecord, IntoRecord, Mutator, Debug, Clone, PartialEq, Eq, Deserialize, PartialOrd, Ord)]
 enum Enum {
     Foo(u32),
     Bar { foo: u64 },
@@ -31,4 +35,15 @@ impl Default for Enum {
     fn default() -> Self {
         Self::Baz
     }
+}
+
+#[derive(FromRecord, IntoRecord, Mutator, Debug, Clone, PartialEq, Deserialize)]
+struct NestedMaps {
+    m: BTreeMap<BTreeMap<u64, String>, BTreeMap<String, BTreeMap<u32, u32>>>
+}
+
+#[derive(FromRecord, IntoRecord, Mutator, Debug, Clone, PartialEq, Deserialize)]
+enum ContainersEnum {
+    Map(BTreeMap<Enum, Struct>),
+    Vec(Vec<NestedMaps>)
 }